TableのAppVersionの更新をとりあえず実装。

+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/Table.kt
@@ -5,7 +5,7 @@ import java.time.Instant
 data class Table(
     val id: TableId,
     val version: Long,
-    val appVersion: Long,
+    val appVersion: Int,
     val hostPlayerId: PlayerId,
     val potManagerPlayerId: PlayerId,
     val rule: Rule,
